requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-cors-headers==3.4.0
django-oauth-toolkit==1.3.2
djangorestframework==3.11.0
idna==2.10
oauthlib==3.1.0
psycopg2==2.8.5
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
